refactor: split verifications/handlers → events + handlers
`handlers.py` previously mixed three concerns: the NPPES pipeline (HTTP,
async), the Claim-cache recompute helpers (nearly pure, no I/O), and the
append-only event writer for non-NPPES §9 transitions. The latter two
are imported by `clinicians/handlers.py` and `org_representations/
handlers.py` without needing any HTTP dependency, making the coupling
unnecessary.

Extract `recompute_clinician_claim`, `recompute_org_claim`, and
`record_verification_event` into `events.py`. Keep the NPPES pipeline
(`run_clinician_verification`, `run_org_verification`, and their admin
retriggers) in `handlers.py`, which now imports the recompute helpers
from `events.py`.

Update all callers and move the pure unit tests for the extracted
functions into `test_events.py`; leave the pipeline tests (NPPES mock
transport, cache write-through) in `test_handlers_phase3.py`.
